live_cook/requirements.txt
fastapi
uvicorn
google-genai
websockets
python-multipart
jinja2
python-dotenv
